api_calendar/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask_cors
flask_marshmallow
python-dotenv